Try making VTXWriter using global arrays

+++ cpp/dolfinx/io/vtk_utils.h
@@ -240,9 +240,7 @@ vtk_mesh_from_space(const fem::FunctionSpace<T>& V)
 /// geometry 'nodes'
 /// @note The indices in the return array correspond to the point
 /// indices in the mesh geometry array
-/// @note Even if the indices are local (int32), both Fides and VTX
-/// require int64 as local input
-std::pair<std::vector<std::int64_t>, std::array<std::size_t, 2>>
+std::pair<std::vector<std::int32_t>, std::array<std::size_t, 2>>
 extract_vtk_connectivity(
     std::experimental::mdspan<const std::int32_t,
                               std::experimental::dextents<std::size_t, 2>>
